+++ sample-android/src/main/kotlin/com/github/panpf/zoomimage/sample/ui/examples/view/PhotoViewFragment.kt
@@ -75,7 +75,6 @@ class PhotoViewFragment : BindingFragment<PhotoViewFragmentBinding>() {
 
     private fun setImage(binding: PhotoViewFragmentBinding) {
         binding.photoView.displayImage(args.imageUri) {
-            lifecycle(viewLifecycleOwner.lifecycle)
             crossfade()
             listener(
                 onStart = {