+++ sample-android/src/main/kotlin/com/github/panpf/zoomimage/sample/ui/examples/view/SketchZoomImageViewFragment.kt
@@ -22,12 +22,13 @@ import androidx.navigation.fragment.navArgs
 import com.github.panpf.assemblyadapter.pager.FragmentItemFactory
 import com.github.panpf.sketch.displayImage
 import com.github.panpf.sketch.resize.Precision
+import com.github.panpf.sketch.stateimage.ThumbnailMemoryCacheStateImage
 import com.github.panpf.zoomimage.ZoomImageView
 import com.github.panpf.zoomimage.sample.databinding.SketchZoomImageViewFragmentBinding
 import com.github.panpf.zoomimage.sample.databinding.ZoomImageViewCommonFragmentBinding
 import com.github.panpf.zoomimage.sample.settingsService
-import com.github.panpf.zoomimage.sample.ui.widget.view.ZoomImageMinimapView
 import com.github.panpf.zoomimage.sample.ui.util.collectWithLifecycle
+import com.github.panpf.zoomimage.sample.ui.widget.view.ZoomImageMinimapView
 
 class SketchZoomImageViewFragment :
     BaseZoomImageViewFragment<SketchZoomImageViewFragmentBinding>() {
@@ -63,8 +64,8 @@ class SketchZoomImageViewFragment :
         onCallError: () -> Unit
     ) {
         binding.sketchZoomImageViewImage.displayImage(args.imageUri) {
-            lifecycle(viewLifecycleOwner.lifecycle)
-            crossfade()
+            placeholder(ThumbnailMemoryCacheStateImage())
+            crossfade(fadeStart = false)
             ignoreExifOrientation(settingsService.ignoreExifOrientation.value)
             listener(
                 onStart = { onCallStart() },
@@ -76,7 +77,6 @@ class SketchZoomImageViewFragment :
 
     override fun loadMinimap(zoomImageMinimapView: ZoomImageMinimapView, sketchImageUri: String) {
         zoomImageMinimapView.displayImage(sketchImageUri) {
-            lifecycle(viewLifecycleOwner.lifecycle)
             crossfade()
             resizeSize(600, 600)
             resizePrecision(Precision.LESS_PIXELS)